Add success and enabled golden tests, extract pin code content in separate widget
